Reworked TrainingSet
Changed Constructor to not expect Patients to be already loaded. Instead, takes a list of their IDs now.
Changed Hashing method for the cache to rely again on
the contained data instead of user-supplied data.
Implemented a method of fetching  often used TrainingSets (Set A, Set B and both) and already instanced training sets based on their name.